+++ subscriptions/subscriptions-impl/src/main/res/layout/legacy_view_itr_settings.xml
